Train_i2v: create file for train_i2v
